Backend/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
google-generativeai
python-dotenv
email-validator
httpx
pytest
pytest-asyncio
pytest-mock